Move `common.mjs` implementation into the `common` folder
Leaves the `common.mjs` file to re-export it so that code importing it from the current release still find it
