+++ src/govuk/components/checkboxes/checkboxes.mjs
@@ -2,7 +2,7 @@ import '../../vendor/polyfills/Function/prototype/bind.mjs'
 // addEventListener, event.target normalization and DOMContentLoaded
 import '../../vendor/polyfills/Event.mjs'
 import '../../vendor/polyfills/Element/prototype/classList.mjs'
-import { nodeListForEach } from '../../common.mjs'
+import { nodeListForEach } from '../../common/index.mjs'
 
 /**
  * Checkboxes component
