+++ src/govuk/components/tabs/tabs.mjs
@@ -3,7 +3,7 @@ import '../../vendor/polyfills/Element/prototype/classList.mjs'
 import '../../vendor/polyfills/Element/prototype/nextElementSibling.mjs'
 import '../../vendor/polyfills/Element/prototype/previousElementSibling.mjs'
 import '../../vendor/polyfills/Event.mjs' // addEventListener and event.target normaliziation
-import { nodeListForEach } from '../../common.mjs'
+import { nodeListForEach } from '../../common/index.mjs'
 
 /**
  * Tabs component
